requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.4
    # via uvicorn
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
click==8.0.1
    # via uvicorn
fastapi==0.65.2
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==2.10
    # via requests
pydantic==1.8.2
    # via fastapi
pyyaml==5.4.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.0
    # via pydantic
urllib3==1.26.5
    # via requests
uvicorn==0.14.0
    # via -r requirements.in
